remove extraneous copy
v3 puts the certificates into FR_EX_TLS_INDEX_CERTS.  v4 doesn't.

so there's no need to create the certs in the wrong context,
then copy them to the correct context, then delete the certs.
We can just create them in the correct context.

Move cert_vps variable into the condition in order to highlight
the fact that it's a temporary variable, used only to limit the
list of certificates being printed.

This change doesn't affect the current issues of EAP-TLS not
caching certs in session state, or it leaking the certs.  But
this change is not wrong.