add section / pair / data in alloc functions, not elsewhere
on review, most of the calls to cf_section_add() were wrong.
the child section was already added to the parent section by
_cf_section_alloc().  There were many, many, calls to
cf_item_add() or cf_section_add() which resulted in the
section being added twice.

However, cf_pair_alloc() and cf_data_alloc() did not add the item
to the parent.

In the interest of less code and fewer errors, we've removed
cf_section_add() and cf_pair_add().  The calls to cf_item_add()
are now done only from the alloc functions.

We still have many calls to cf_data_add(), because that function
is in fact allocating *and* adding the CONF_DATA items.