more cleanups
remove "load dictionary" code from bootstrap phase.
Replace it with an assert that the dictionary exists.

move the "compile" function back to the instantiate phase.
it was erroneously moved to the bootstrap phase a few commits ago.

if there is a compile function, assert that the dictionary it uses
is the same as the dictionary used for the namespace.

load the compiler function by namespace, not by server second
name.