cross-check name / number for attributes
we previously looked up (and checked) duplicate attributes by name.

We now check if attributes of the same number are compatible.
If not, error out.

As a result, a number of dictionaries have to be updated in order
to have compatible types.  Mostly by commenting out one of the
incompatible types.