add "is_dup" flag
so if we have 2 attributes of the same number / type, but different
names.  We mark the old one as "dup".

Then when looking up the attribute by name, check for the dup
flag, and if found, return the *new* attribute.

This change lets us decode packets into the new names, then
read config files using the old names, and have the two
attributes match.