Make rbtree_walk() users use iterators instead.
That makes rbtree_walk() unused, so it goes away, too.
In turn, since rbtree_walk() is the only caller of the
walk_*_order() functions, they go away as well, and it
means the various "compare" functions (that do what's
usually called "visiting" nodes in the context of tree
traversal) will typically be inlined into the one place
they're used, in the iterator loop, and control flow
that's implied by what the "compare" functions return
can just be explicitly done in said loop.