reset request->module to NULL
when unlang calls the process function, it thinks that it's
calling a module, so it sets request->module = name of process
function.

This means that all debug messages (such as "received packet")
are printed out, prefixed with the process name.  This makes
debug messages harder to read.

Since the first thing the process function does is call an unlang
section, we can safely set request->module=NULL, which omits the
"radius - " prefix from the packet we're printing