clean up
use "da" instead of "child", as it's more consistent.

Don't clamp "data_len", as it should be used when we see
"end of options" signifier, to indicate that the decode function
is eating all of the remaining data.

and, all of subsequent calls to various functions use data[1]
as the length, and ignore data_len