Move meaning of tls13_enable to TTLS and PEAP
Now that we always allow TLS 1.3 for RadSec and EAP-TLS, if
(a) OpenSSL supports it, and (b) it's configured in tls_max_version

We don't allow TLS 1.3 for TTLS and PEAP, unless a magic flag
is set.

We don't allow TLS 1.3 for FAST, as we don't know what to do there.