clean up and de-dup FR_DBUFF_MAX_NO_ADVANCE()
which was being called multiple times, depending on the call flow.

It now just passes the input buffers through, and any function
which knows it can only encode X bytes will do

	work_dbuff = FR_DBUFF_MAX_NO_ADVANCE(dbuff, X)

There may also be issues with dbuffs, as the encode_extended()
was being called _always_ with max 255 bytes.  Yet it somehow
managed to encode more than 255 bytes in the tests