don't access list after it's been freed
and don't free a list on parse error.  We rely instead on the
caller to free things.

If we do want to free the temporary lists while reading, we will
have to add a temporary context, OR walk down the temporary list
freeing each individual element.

The temporary lists are allocated on the stack, so we cannot
talloc_free() them.