Use calloc(), not malloc() to allocate heap/lst elements from the (memory) heap (#4175)
malloc(), in the C standard and in the man page, is described as not
initializing the memory it allocates. A standard-conforming C compiler
could thus compile versions of the tests that fail fr_{heap, lst}_insert()
checks that keep an item from being inserted twice or into more than one
heap or lst with the same index displacement.

calloc() zeroes the allocated memory.