allow dns label decoding to stop at the 0x00 byte
so that when we're decoding DNS labels in a struct, or in an
array, we can stop when the labels end.  The next field is then
whatever happens to be after the DNS labels.

However, when they're encoded in an option, they must exactly
fill the option, otherwise it's an error