track and clean up old packets
When we send a reply, track the State in a new hash table,
and associate the fr_io_track_t with that state.

When we create a new tracking entry for a packet with a State,
look up the &request.State.  If there's a matching entry, then
free it.  Which means that we don't keep old packets around
in the dedup tree.