Define fr_ldap_referral_t to hold parsed LDAP referrals while they are followed
A given LDAP resonse which contains referral details can consist of
multiple referral URLs all of which should be capable of returning the
same data.

We parse all of the URLs and firstly attempt to query a server that
already has an active trunk connection.

All the possible referral URLs may need to be parsed to find an active
connection.

If none of the referral URLs point to an active connection, then the
parsed data stored in this structure is used to create new connections,
rather than re-parsing.