add fr_time_scale() and use it when parsing times.
so that we can clamp the times at min/max values, if the user
gives us something which is out of bounds.

we still need to fix value_box_from_network for DATE, as those
values are unsigned.  And also fix the various casting functions
to clamp the relevant values.