create and use T_ADD, T_SUB, T_MUL, and T_DIV
note that we don't (yet) add them to fr_tokens_table[]

That's because attribute names have "-" in them.  Adding "-" as
a token means that we no longer parse attribute names in the
same way.

The solution is likely to update the various pair parsers to
avoid the 1999-era gettoken() APIs, and instead use more modern
equivalents.

For now, we just define and use add/sub/mul/div for the various
calc tests.