minor cleanups, and add calculations for booleans
it's still not clear what to do for IFIDs and ethernet addresses.
IFIDs are site-local integers, and likely should just be treated
as opaque blobs.

we may want to allow editing ethernet addresses, by splitting
them into OUI, or by merging OUI and trailing bits.  But for now
the caller might as well just convert them to octets, do the
opertions, and convert them back.