Revert "reorder types"
There are so many uses of the old ordering it makes visually checking whether cases are covered annoying.

The old ordering orders things by types which relate directly to C types, and then more abstract types, and system types, which is generally far more useful.

The C types are usually the ones we deal with explicitly in martialling and unmartialling functions.  It's very rare we end up needing to convert between time_delta and a native database type, but very common we need to deal with the numeric and float types.