Add xctx structures for xlat instantiation and calling
This is similar to the changes made to modules.

These changes also switch the instance and thread instance trees to heaps, which allows O(1) lookup of thread instance data.  This is significantly more efficient than the old code.

This commit is large as it changes the function signature for xlat functions, and removes a large amount of boilerplate in the modules that dealt with copying module instance and thread instance data around.