fr_pair_list_afrom_str() now takes a dict_attr root, not dict
in preparation for allowing structural attributes to be parsed
from strings.

This helps with not only intermediate tests while the new edit
code is being developed, it allows for pairs to be assigned
as the result of xlat / exec expansions.  Which gives a fair bit
more generality to the server.