tell the user things that OpenSSL won't.
If they try to allow TLS versions which aren't supported by
the current security level, then tell the user to change the
security level.

By default, OpenSSL will just ignore the TLS versions, and not
tell the user anything.

Instead of leaving the user to figure things out, we give a
helpful error.  Because we don't hate our users.