don't try to open a new socket if we're full
i.e. instead of "open, add, fail, close", it's "try to open, fail"

which avoids work and potential race conditions.

The downside is that there is a small race condition where sometimes
it will either open (or not) new connections when the list is almost
full.  Either situation is fine, and should be handled by the code.

The important thing is to give the admin an earlier error, and
a more descriptive error.