add "ignore this list" entries.
If we're inserting a new structural member, then we don't need
to undo edits to any children in it's lists.  Those edits will be
undone automatically when the parent pair is undone.

We keep a separate list of "lists to ignore", and check that
before doing anything.  If the list can be ignored, then we
discard delete / clear / etc.

In inserts, if the VP is structural, then remember its child list,
so that we can ignore edits to it, too.