Always put something in temporary buffer, if only "" (CID #1503992)
From the point where src->dst_type is determined to have fixed
size, fr_value_box_from_substr() works with a local copy... but
if fr_sbuff_remaining() returned zero, nothing was written in the
local copy. With this change, in that case, the local copy is made
an empty string, NUL-terminated.