double-check and fix flags.
We now have a "can_purify" flag.  This is set to "true" for
pure functions which have pure arguments.  And for alternations
which have pure arguments.

%% is set to pure.  XLAT_BOX and XLAT_LITERAL are pure, but
they are not "can_purify".

The result is that we tell if there are portions of the tree
which need additional work for purification.