add --enable-fips-workaround flag to configure
which means that we use our internal MD4 / MD5 functions always,
as OpenSSL requires much magic to re-enable them in FIPS mode.

It's easier to just our internal implementations, instead of
butchering the stable v3 release.

The downside is that if you do a default build and then enable
FIPS in OpenSSL, the server will simply not work.  It has to be
re-build with the --enable-fips flag.