allow for instantiated xlats to be put into other ones
tmpl_tokenize() will call xlat_instantiate_ephemeral() on things
it sees.  However, if we then hoist that xlat out of the tmpl
and into the main xlat tree, we can't re-instantiate the node.

Rather than adding flags to change the rest of the code, we just
catch that here.  This should likely be cleaned up later