don't expand structural types when used in "truthy" context
	&reply || ...

should NOT get a list of value-boxes out of the reply.  Instead,
it should just check for the "truthiness" of &reply.

We still need to add a special case for when a list is used by
itself, as in

	if (&reply) { ...

but that can be handled with a few lines of code in the condition
evaluation code.