Annotate issues Coverity see that we have decided to ignore.
This change deals with the following CIDs:
1503889, 1503901, 1503908, 1503916, 1503936, 1504024, 1504280,
1504282, 1503894, 1503899, 1504013

Most of these are sbuff/dbuff operations that can only fail if
the buffers can't be reallocated or have reached their specified
maximum size.

In 1503899, we know, but Coverity doesn't, that no log_keyword
passed to _make_vlog() will overrun the fixed-size buffer it uses.

1503894 is what Coverity thinks is a copy/paste error that clearly
makes sense as written.