Annotate calls that as used will not return NULL (CID #1503961) (#4577)
The calls in question are those in pair_list_init() that iterate
over a pair list built up from a string. As long as the string
is non-empty, the pair list won't be empty. As long as perc (the
percentage) is no larger than 100, the code won't pass NULL to
fr_pair_copy()... and one can see that's the case for all calls
to pair_list_init(). Coverity apparently doesn't look at that
context.