horrible temporary hacks to parse conditions
once the config file stuff is moved to sbuffs, this can be deleted.

The root cause of this hack is to avoid having the xlat code call
cf_expand_variables().  The old condition code did that, which "worked"
but was arguably wrong.

Moving the cf_expand_variables() call to cf_file.c made it difficult
to parse multi-line "if" conditions.  So we necessitate this hack.