Uniform way of dealing with static analysis
Static analysis doesn't necessarily infer all properties of code
and thus may give false positives. Since we use at least two static
analyzers, clang's and coverity, we define STATIC_ANALYZER if and
only if some static analysis is running, and check it instead of
a macro specific to a particular static analyzer. It's defined,
or not, in src/include/build.h, so that adding a new static
analyzer only requires adding code, not changing it.