add flatten / unflatten API
which convert pairs from a (mostly) flat list of leaves, to
a tree-like structure.

Many of the protocols alread decode/encode TLVs, so those are
left alone.  Groups are left alone, as DHCPv6 uses them a lot.

structs likely need more tests to be sure that they work in all
cases, including when structs contain structs