Attempt to quiet coverity on fr_value_box_t assignments (CIDs follow)
CIDs: 1508477, 1508479, 1508480, 1508481, 1508482, 1508483,
      1508484, 1508486

Some members of an fr_value_box_t are const qualified, hence the
complaints about attempts to assign to them or to members of the
union of possible values, some of which are const qualified
(though coverity doesn't appear to complain consistently, e.g. the
FR_TYPE_UINT8 and FR_TYPE_UINT16 cases in the switch in
sim_decode_pair_value() in src/lib/eap_aka_sim/decode.c). We're
trying not to use annotations.