if we set use_new_conditions, don't even parse the old ones
and remove the one test which no longer fails at load time.
The old code would check if the condition could at all be done,
such as assigning a value which is too large.

The old code would see that one side had a fixed data type, and
try to cast the other side to that data type.  The new code does not
(as yet) do this