Annotate calls we know won't fail (CID #1503939)
The individual tests in edit_tests all operate on a three-element
pair list they get from add_pairs(), freeing it after they're done.
fr_pair_afrom_da() only returns NULL if it runs out of memory--
unlikely in this context, so we annotate the calls that add items
to the list.