Annotate "tainted data" in fr_udp_header_check() call (CID #1504049) (#4727)
Coverity claims that the "downcast" of a uint8_t const * to
ip_header_t const * taints "the data that this pointer points to".
"this pointer" must be p, because coverity says not a word about
dereferences of ip. The pointed at data isn't changed, being const
qualified. p and the passed length are checked in fr_udp_header_check().