run aclocal on make reconfig
aclocal will automatically pull in extra macros from the system
macro dir if not defined locally, and put everything in aclocal.m4

We do it for all modules as well, so autoconf no longer needs a
search path for macros.

aclocal crashes out if the include dir doesn't exist, but we are moving
all macros into the top-level m4/ directory, so just stop looking in a
local per-module m4/ directory.

Also remove the top-level aclocal.m4, because we will now
regenerate it when needed.