radeapol_uat: Add new script for user tests
We need something similar to scripts/util/raduat, but instead of using radclient. it needs to be implemented in python calling the eapol_test.

the python one needs to also take an eapol_test config as part of the request
as a separate file like:

"my_peap_test", "my_peap_test_conf", "my_peap_test_expected"

"my_peap_test" contains all the attributes we want to send
"my_peap_test_conf" contains the eapol test config
"my_peap_test_expected' contains the attributes we want to see , and whether it'll be an Access-Accept or Access-Reject

the python script needs to implement something similar to raduat, with how the test files are laid out
it needs to take "my_peap_test", "my_peap_test_conf", and figure out how to pass them to eapol_test (maybe using the radict)

i.e. how to get the attributes out of "my_peap_test" and pass them in a format eapol_test wants
passing the config and when eapol_test finishes, checking to see if the attributes it returned match "my_peap_test_expected"