Ignore returns from certain sbuff calls (CIDs #1504025, #1504070) (#4750)
These are cases in which sbuffs are either created from C arrays
or are allocated from the heap but given initial and maximum sizes.
The author(s) of the code presumably chose sizes intended to suffice,
but despite that, coverity will notice that the sbuff functions
used return a value, and hence will report check_return barring
either explicit cast to void or annotation. We therefore cast to void.