Fix the use of the ctype classification and mapping functions.
Even though the argument type for these functions is `int`, there
is only a restricted subset of values which are actually valid.
The values are either EOF (which is negative) or the values
representable as an `unsigned char`, says the standard.

On NetBSD with pkgsrc, where we typically build with -Wchar-subscripts,
the existing use (without casts of arguments) typically gets us
   warning: array subscript has type 'char'

Someone has apparently been in here before to silence warnings and
added casts to int, which will invoke undefined behaviour for half
of the values of a `signed char` argument, as a negative `signed
char` will be sign-extended to a negative `int` by the cast, and
most probably not to the EOF value.

I could not offhand see any use which would deal with EOF, so added
casts to `u_char` to make the behaviour well-defined, and this
should both properly get rid of the warnings and the undefined
behaviour.