Add -Wdeclaration-after-statement where supported (#4762)
* Add -Wdeclaration-after-statement where supported

This enforces one of the FreeRADIUS coding standards.
NOTE: I had to do a "make force-reconfig" to get the changes
made to configure.ac and m4/ax_cc.m4 to have an effect in
configure. This caused changes in thirty-five other configure
files that are unrelated to this change. I am including only
the changes to the files relevant to adding the above option
to CFLAGS, namely the files I explicitly changed and configure
itself.

We also add changes to fix the declaration after statement
instances this addition turns up.

* Don't declare vars in loops

* Don't declare vars in groups

Co-authored-by: Arran Cudbard-Bell <a.cudbardb@freeradius.org>