Don't compare with bounds in SBUFF_PARSE_[U]INT_DEF (CIDs 15040045, 1504004)
Explicit comparison with the bounds makes sense but when you hit
the type the underlying conversion function uses, it's pointless,
and when it's part of a larger expression, coverity considers it
a possible coding error.

So... rather than doing an explicit comparison, we use the following
characterization: it's an error if the cast to the desired type changes
the value. To distinguish between overflow and underflow for signed
integer types, overflow happens if the cast value is less than the
return from strtoll(), underflow if it's greater.