forbid array references in sublists
	&list := {
		&foo[1] = "bar"
	}

The edit code does not support this, because it doesn't really
make a lot of sense.  If we do need to support it, then the edit
code needs to be updated to support it.

Note that we can already do

	&list.foo[1] = "bar"

so this limitation isn't much of an issue.  The problem comes when
doing multiple attribute assignments, and those using array indexes
inside of a sublist.  What does it mean when we do

	&list := {
		&foo[0] = "bar"
		&foo[1] = "baz"
		&foo[0] = "ugh"
	}

What the heck is going on here?  The real answer is "don't do that"