preliminary support for compiling local variables
However, it doesn't quite work as yet.  The tmpl_rules_t have
parents, and this a parent dict to check.  But the pass2 resolution
functions just get passed the local dictionary, and not the entire
tmpl_rules_t.

As a result, the attributes are looked up in either the local
dictionary. or in the protocol one.  But there's no fallback
from one to the other.