use enums when upcasting for comparisons
Note that (for now), we're not doing it in tmpl_cast_in_place(),
which *explicitly* avoids passing enums when casting.

We're also not doing it for expression / math calculations, as
it doesn't make sense to do

	Service-Type = Framed-User + 1

It might make sense to use enums for IP address calculations? e.g.

	Framed-IP-Address = network & mask

but at that point, you might as well just use attributes or local
variables for everything.

This is in preparation for updating the parser to force the user
to use single-quoted strings for enum names, as in:

	Service-Type == 'Framed-User'

Which then means we'll also have to update the printing routines
to always print quotation marks around enums, and then update the
users file, SQL, detail file, etc.