add functions to get gmt offset and daylight savings flag
we should note that the server tracks times in UTC, so

	(date) foo % (time_delta) 1d

will return the offset from the UTC day, not from the local time.
This could perhaps be more prominent.  But tracking times internally
in UTC is the only way to make sure we're not screwed up by time
zone and DST changes.