fix "success" policy to use new edit sections
we don't need to do "if" checks for parent, we can just use
the "=" operator, and let that fail-safe.

We also revert the "subrequest" test to the previous version.

The next step is to figure out why changing the xlat expansions
broke the "success" policy.  But we can check that tomorrow,
once the CI tests pass again.