Allow using clang 15 with -Werror and -Wunreachable-code-generic-assoc
Pre C17, there was ambiguity about the type of the controlling expression
in __Generic(). C17 explicitly states that it is subject to rvalue
promotions, which removes qualification from the outermost (or is that
deepest?) type modifier (or the base type if there are no modifiers).

(To clarify with an example, a FOO const * const will become FOO const *
as far as __Generic() is concerned.)

clang 15, perhaps anticipating C17, has -Wunreachabale-code-generic-assoc,
which warns if a __Generic() alternative type is unreachable, e.g.
FOO const * const. This change gets rid of such alternatives in a couple
of macros.

This builds and passes such tests as I can do compiling with gcc, clang 14,
and clang 15.