don't mangle request->session_state_ctx manually
now that it's an actual list, and in the pair_root, we need to
remove it from the pair_root if it's freed, and add it to the
pair_root when it's allocated or restored.

and a check of

	git grep -- 'request->session_state_ctx = '

should return only one assignment in request_state_replace().
Nothing else should muck with the state