remove state tracking
the configuration isn't exposed, and even if it was, the code
doesn't do anything other than track state.  Nothing uses that
state for anything.

If we want to use State to direct EAP packets to the same back-end,
the correct place to put that is likely in the fr_io_track_t
structure.  Which has to be written to in the "encode" or "write"
routine, and then also in the "read" routine.

The network side can then use that field to have a centralized
tracking structure of state -> worker.  Even if this state
is probabilistic, at 64 bits (and a reasonable hash function),
the odds of a packet going wrong are tiny.  If it's an issue,
we could just change the state to a talloc ptr, and use memcmp()
to compare states.

That also lets us put anything into the state, which is flexible.