Upload the result of preprocessing a modeling file
We can't run cov-make-library ourselves, which is the only command
that lets you say where to find header files. So, we'll preprocess
a single model file that includes whatever we choose to model the
stubs for, grouped at least by the source file they come from, each
group starting with #include directives appearing in the source file.