allow setting SINGLE_CONNECT mode.
note that we *or* in the SINGLE_CONNECT flag, because the other
flags control packet encryption.  And we don't want the admin to
magically send unencrypted packets when there's a shared secret.

We also do this only when sending replies from the server.  If
we're encoding a client packet, the "flags" attribute must already
exist, and must already have been encoded using fr_struct_to_network()