remove assertion
We need to fix up encode_tlv(), as for RADIUS, it's really
"encode TLV contents", and not "encode the entire TLV".

The encode_child() function also relies on the encode_value()
funtion to handle grouping attributes, which is likely also wrong.

The RADIUS encoder / decoder is not only more complex than the other
protocols, it was written earlier.  i.e. before we cleaned up and
clarified the process of encoding packets.