Separate out python2/3 packages for Debian systems
Until now we built whatever the default Python was on the system,
and not build the other module - this was included in the
freeradius package.

Separating it out makes it easier for people to choose which they
want. Recommends dependency on the freeradius package because we
previously bundled the module with that.