separate out NAS-Port from NAS-Port-ID.  Helps with #5010
radutmp and couchbase use NAS-Port for their fields, but SQL uses
NAS-Port-Id.  Which is a string, not an integer.

So update the API to allow for both to be passed.  If there's
NAS-Port-Id, we use it.  Otherwise we use NAS-Port.  So radutmp
and couchbase don't change, and SQL now uses the correct field.

This is a behavior change from earlier.  However, it means that
we are dropping an incorrect NAS-Port, and replacing it with a
NAS-Port-Id.

The only real conclusion here is that almost no one uses simultaneous
use with SQL any more.